auto jira

+++ tests/conftest.py
@@ -33,12 +33,12 @@ def after_each_test_run(request):
             continue
         # First, delete all tests associate with this issue key on TSM.
         tests = test_service.get_tests_in_issue(issue_key)
-        for _, test_key in tests:
-            test_service.delete_test(test_key)
-        # Or first, delete all tests have name exist
         # for _, test_key in tests:
-        #     if (_ == _tests[issue_key][0]['name']):
-        #         test_service.delete_test(test_key)
+        #     test_service.delete_test(test_key)
+        # Or first, delete all tests have name exist
+        for _, test_key in tests:
+            if (_ == _tests[issue_key][0]['name']):
+                test_service.delete_test(test_key)
 
         # Second, for each tests have found by pytest,
         # associate with this issue key, create and update its testCaseKey.
@@ -46,12 +46,18 @@ def after_each_test_run(request):
             test_key = test_service.create_test(issue_key, test)
             test['testCaseKey'] = test_key
 
-        # Last, create a test cycle for current issue key on TSM.
+        # Third, create a test cycle for current issue key on TSM.
         cycle_items = [{
             key: item[key]
             for key in ['testCaseKey', 'status']
         } for item in _tests[issue_key]]
         test_service.create_test_cycle(issue_key, issue_key, cycle_items)
+        # Last, transition status of task is Production if status of task is Ready for TEST
+        for j in range(len(_tests[issue_key])):
+            if 'Fail' in _tests[issue_key][j]['status']:
+                break
+            elif issue['fields']['status']['name'] == 'Ready for TEST':
+                test_service.transition_status_task(issue_key)
 
 
 @pytest.fixture(autouse=True)
